Bug #1727931 sched: Introduce task-rotation feature
[root cause  ] improve scheduler
[changes     ] introduce task-rotation feature
[side effects] performance
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:yes
[board]:sharkl5pro
[test case]:geekbench scheduler
[reviewers   ] ke.wang vincent.wang
[change_type ] feature
[tag_product ] common

In ARM Big.Little cpu topology, big task's number may be more than
the number of cpu. In this case, some big tasks may only could run
on the little core, it is not fair and performance maybe loss.

Introduce rotation feature to improve the performance.

At scheduler_tick, it would check the big core's task and small core's task,
when the small's task is misfit, and the big core's task had running
for a long time. It would swap the two task's cpu, so that the small core's
task can run on big core.

As a result, the big tasks can rotation running between little core and
big core to improve the performance.

Below is the test in Geekbench:

                origin  modify
muiti-core:     1460    1581(+8.3%)
crypto:         2583    2691(+4.2%)
integer:        1431    1560(+9.0%)
floating:       1338    1443(+7.8%)

Change-Id: Icfaafab7b3ea95b8143464bba8bdf83d82d312b8
Signed-off-by: Xuewen Yan <xuewen.yan@unisoc.com>
[dereference23: Kill check_for_migration()]
[Pzqqt: Forward ported to android12-5.10]
Signed-off-by: Alexander Winkowski <dereference23@outlook.com>